llmservice-main/requirements.txt
# 파일 생성 완료
# 의존성 설정 

#LLM & AI 
google-genai>=1.0.0

#Vector Database 
chromadb>=0.4.0

#Database 
#설치 방법 : pip install mysql-connector-python
mysql-connector-python>=8.0.0

#HTTP Requests(for API calls)
requests>=2.31.0

#Utillities
python-dotenv>=1.0.0

